+++ bindings/ruby/lib/stateset_embedded.rb
@@ -7,5 +7,5 @@ rescue LoadError
 end
 
 module StateSet
-  VERSION = '0.9.9'
+  VERSION = '1.0.0'
 end