+++ bindings/ruby/spec/commerce_spec.rb
@@ -317,6 +317,6 @@ end
 
 RSpec.describe StateSet do
   it 'has a version number' do
-    expect(StateSet::VERSION).to eq('0.9.9')
+    expect(StateSet::VERSION).to eq('1.0.0')
   end
 end